# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"_36_Enums.TipoUsuario": {
				"type": "string",
				"enum": [
					"INDIVIDUAL",
					"EMPRESARIAL"
				]
			},
...
			"_36_Enums.rolEquipo": {
				"type": "string",
				"enum": [
					"LECTOR",
					"COMENTARISTA",
					"EDITOR"
				]
			},
			"EquipoDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"correo": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"nombreCompleto": {
						"type": "string"
					},
					"telefono": {
						"type": "string"
					},
					"direccion": {
						"type": "string"
					},
					"fotoPerfil": {
						"type": "string"
					},
					"rolEquipo": {
						"type": "string",
						"enum": [
							"LECTOR",
							"COMENTARISTA",
							"EDITOR"
						]
					},
					"estado": {
						"type": "string",
						"enum": [
							"activo",
							"inactivo"
						]
					},
					"empresaId": {
						"type": "number",
						"format": "double"
					},
					"perfilCompleto": {
						"type": "boolean"
					}
				},
				"required": [
					"username",
					"correo",
					"password",
					"nombreCompleto",
					"telefono",
					"direccion",
					"rolEquipo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_EquipoDTO_": {
				"properties": {
					"username": {
						"type": "string"
					},
					"correo": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"nombreCompleto": {
						"type": "string"
					},
					"telefono": {
						"type": "string"
					},
					"direccion": {
						"type": "string"
					},
					"fotoPerfil": {
						"type": "string"
					},
					"rolEquipo": {
						"type": "string",
						"enum": [
							"LECTOR",
							"COMENTARISTA",
							"EDITOR"
						]
					},
					"estado": {
						"type": "string",
						"enum": [
							"activo",
							"inactivo"
						]
					},
					"empresaId": {
						"type": "number",
						"format": "double"
					},
					"perfilCompleto": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"legacyJwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "producttrack",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "@mariaximena0724 y emilia"
		}
	},
	"paths": {
...
		"/equipo": {
			"post": {
				"operationId": "CrearEquipo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"deletedAt": {
													"type": "string",
													"format": "date-time"
												},
												"idUsuario": {
													"type": "number",
													"format": "double"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"perfilCompleto": {
													"type": "boolean"
												},
												"empresaId": {
													"type": "number",
													"format": "double"
												},
												"estado": {
													"type": "string"
												},
												"rolEquipo": {
													"$ref": "#/components/schemas/_36_Enums.rolEquipo"
												},
												"rol": {
													"type": "string"
												},
												"nit": {
													"type": "string"
												},
												"nombreEmpresa": {
													"type": "string"
												},
												"tipoUsuario": {
													"$ref": "#/components/schemas/_36_Enums.TipoUsuario"
												},
												"fotoPerfil": {
													"type": "string"
												},
												"direccion": {
													"type": "string"
												},
												"telefono": {
													"type": "string"
												},
												"nombreCompleto": {
													"type": "string"
												},
												"password": {
													"type": "string"
												},
												"correo": {
													"type": "string"
												},
												"username": {
													"type": "string"
												}
											},
											"required": [
												"deletedAt",
												"idUsuario",
												"updatedAt",
												"createdAt",
												"perfilCompleto",
												"empresaId",
												"estado",
												"rolEquipo",
												"rol",
												"nit",
												"nombreEmpresa",
												"tipoUsuario",
												"fotoPerfil",
												"direccion",
												"telefono",
												"nombreCompleto",
												"password",
												"correo",
												"username"
											],
											"type": "object"
										},
										{
											"properties": {
												"mensaje": {
													"type": "string"
												}
											},
											"required": [
												"mensaje"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Equipo"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EquipoDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ObtenerTodosLosEquipos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"properties": {
													"deletedAt": {
														"type": "string",
														"format": "date-time"
													},
													"idUsuario": {
														"type": "number",
														"format": "double"
													},
													"updatedAt": {
														"type": "string",
														"format": "date-time"
													},
													"createdAt": {
														"type": "string",
														"format": "date-time"
													},
													"perfilCompleto": {
														"type": "boolean"
													},
													"empresaId": {
														"type": "number",
														"format": "double"
													},
													"estado": {
														"type": "string"
													},
													"rolEquipo": {
														"$ref": "#/components/schemas/_36_Enums.rolEquipo"
													},
													"rol": {
														"type": "string"
													},
													"nit": {
														"type": "string"
													},
													"nombreEmpresa": {
														"type": "string"
													},
													"tipoUsuario": {
														"$ref": "#/components/schemas/_36_Enums.TipoUsuario"
													},
													"fotoPerfil": {
														"type": "string"
													},
													"direccion": {
														"type": "string"
													},
													"telefono": {
														"type": "string"
													},
													"nombreCompleto": {
														"type": "string"
													},
													"password": {
														"type": "string"
													},
													"correo": {
														"type": "string"
													},
													"username": {
														"type": "string"
													}
												},
												"required": [
													"deletedAt",
													"idUsuario",
													"updatedAt",
													"createdAt",
													"perfilCompleto",
													"empresaId",
													"estado",
													"rolEquipo",
													"rol",
													"nit",
													"nombreEmpresa",
													"tipoUsuario",
													"fotoPerfil",
													"direccion",
													"telefono",
													"nombreCompleto",
													"password",
													"correo",
													"username"
												],
												"type": "object"
											},
											"type": "array"
										},
										{
											"properties": {
												"mensaje": {
													"type": "string"
												}
											},
											"required": [
												"mensaje"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Equipo"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/equipo/filtrar": {
			"get": {
				"operationId": "FiltrarEquipos",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"properties": {
													"deletedAt": {
														"type": "string",
														"format": "date-time"
													},
													"idUsuario": {
														"type": "number",
														"format": "double"
													},
													"updatedAt": {
														"type": "string",
														"format": "date-time"
													},
													"createdAt": {
														"type": "string",
														"format": "date-time"
													},
													"perfilCompleto": {
														"type": "boolean"
													},
													"empresaId": {
														"type": "number",
														"format": "double"
													},
													"estado": {
														"type": "string"
													},
													"rolEquipo": {
														"$ref": "#/components/schemas/_36_Enums.rolEquipo"
													},
													"rol": {
														"type": "string"
													},
													"nit": {
														"type": "string"
													},
													"nombreEmpresa": {
														"type": "string"
													},
													"tipoUsuario": {
														"$ref": "#/components/schemas/_36_Enums.TipoUsuario"
													},
													"fotoPerfil": {
														"type": "string"
													},
													"direccion": {
														"type": "string"
													},
													"telefono": {
														"type": "string"
													},
													"nombreCompleto": {
														"type": "string"
													},
													"password": {
														"type": "string"
													},
													"correo": {
														"type": "string"
													},
													"username": {
														"type": "string"
													}
												},
												"required": [
													"deletedAt",
													"idUsuario",
													"updatedAt",
													"createdAt",
													"perfilCompleto",
													"empresaId",
													"estado",
													"rolEquipo",
													"rol",
													"nit",
													"nombreEmpresa",
													"tipoUsuario",
													"fotoPerfil",
													"direccion",
													"telefono",
													"nombreCompleto",
													"password",
													"correo",
													"username"
												],
												"type": "object"
											},
											"type": "array"
										},
										{
											"properties": {
												"mensaje": {
													"type": "string"
												}
											},
											"required": [
												"mensaje"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Equipo"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "nombreCompleto",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "correo",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "rolEquipo",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"LECTOR",
								"COMENTARISTA",
								"EDITOR"
							]
						}
					}
				]
			}
		},
		"/equipo/{id}": {
			"get": {
				"operationId": "ObtenerEquipoPorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"deletedAt": {
													"type": "string",
													"format": "date-time"
												},
												"idUsuario": {
													"type": "number",
													"format": "double"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"perfilCompleto": {
													"type": "boolean"
												},
												"empresaId": {
													"type": "number",
													"format": "double"
												},
												"estado": {
													"type": "string"
												},
												"rolEquipo": {
													"$ref": "#/components/schemas/_36_Enums.rolEquipo"
												},
												"rol": {
													"type": "string"
												},
												"nit": {
													"type": "string"
												},
												"nombreEmpresa": {
													"type": "string"
												},
												"tipoUsuario": {
													"$ref": "#/components/schemas/_36_Enums.TipoUsuario"
												},
												"fotoPerfil": {
													"type": "string"
												},
												"direccion": {
													"type": "string"
												},
												"telefono": {
													"type": "string"
												},
												"nombreCompleto": {
													"type": "string"
												},
												"password": {
													"type": "string"
												},
												"correo": {
													"type": "string"
												},
												"username": {
													"type": "string"
												}
											},
											"required": [
												"deletedAt",
												"idUsuario",
												"updatedAt",
												"createdAt",
												"perfilCompleto",
												"empresaId",
												"estado",
												"rolEquipo",
												"rol",
												"nit",
												"nombreEmpresa",
												"tipoUsuario",
												"fotoPerfil",
												"direccion",
												"telefono",
												"nombreCompleto",
												"password",
												"correo",
												"username"
											],
											"type": "object"
										},
										{
											"properties": {
												"mensaje": {
													"type": "string"
												}
											},
											"required": [
												"mensaje"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Equipo"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "ActualizarEquipo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"deletedAt": {
													"type": "string",
													"format": "date-time"
												},
												"idUsuario": {
													"type": "number",
													"format": "double"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"perfilCompleto": {
													"type": "boolean"
												},
												"empresaId": {
													"type": "number",
													"format": "double"
												},
												"estado": {
													"type": "string"
												},
												"rolEquipo": {
													"$ref": "#/components/schemas/_36_Enums.rolEquipo"
												},
												"rol": {
													"type": "string"
												},
												"nit": {
													"type": "string"
												},
												"nombreEmpresa": {
													"type": "string"
												},
												"tipoUsuario": {
													"$ref": "#/components/schemas/_36_Enums.TipoUsuario"
												},
												"fotoPerfil": {
													"type": "string"
												},
												"direccion": {
													"type": "string"
												},
												"telefono": {
													"type": "string"
												},
												"nombreCompleto": {
													"type": "string"
												},
												"password": {
													"type": "string"
												},
												"correo": {
													"type": "string"
												},
												"username": {
													"type": "string"
												}
											},
											"required": [
												"deletedAt",
												"idUsuario",
												"updatedAt",
												"createdAt",
												"perfilCompleto",
												"empresaId",
												"estado",
												"rolEquipo",
												"rol",
												"nit",
												"nombreEmpresa",
												"tipoUsuario",
												"fotoPerfil",
												"direccion",
												"telefono",
												"nombreCompleto",
												"password",
												"correo",
												"username"
											],
											"type": "object"
										},
										{
											"properties": {
												"mensaje": {
													"type": "string"
												}
											},
											"required": [
												"mensaje"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Equipo"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_EquipoDTO_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "EliminarEquipo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"deletedAt": {
													"type": "string",
													"format": "date-time"
												},
												"idUsuario": {
													"type": "number",
													"format": "double"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"perfilCompleto": {
													"type": "boolean"
												},
												"empresaId": {
													"type": "number",
													"format": "double"
												},
												"estado": {
													"type": "string"
												},
												"rolEquipo": {
													"$ref": "#/components/schemas/_36_Enums.rolEquipo"
												},
												"rol": {
													"type": "string"
												},
												"nit": {
													"type": "string"
												},
												"nombreEmpresa": {
													"type": "string"
												},
												"tipoUsuario": {
													"$ref": "#/components/schemas/_36_Enums.TipoUsuario"
												},
												"fotoPerfil": {
													"type": "string"
												},
												"direccion": {
													"type": "string"
												},
												"telefono": {
													"type": "string"
												},
												"nombreCompleto": {
													"type": "string"
												},
												"password": {
													"type": "string"
												},
												"correo": {
													"type": "string"
												},
												"username": {
													"type": "string"
												}
											},
											"required": [
												"deletedAt",
												"idUsuario",
												"updatedAt",
												"createdAt",
												"perfilCompleto",
												"empresaId",
												"estado",
												"rolEquipo",
												"rol",
												"nit",
												"nombreEmpresa",
												"tipoUsuario",
												"fotoPerfil",
												"direccion",
												"telefono",
												"nombreCompleto",
												"password",
												"correo",
												"username"
											],
											"type": "object"
										},
										{
											"properties": {
												"mensaje": {
													"type": "string"
												}
											},
											"required": [
												"mensaje"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Equipo"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/equipo/forzar-eliminar/{id}": {
			"delete": {
				"operationId": "EliminarFisicamente",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"eliminado": {
													"properties": {
														"deletedAt": {
															"type": "string",
															"format": "date-time"
														},
														"idUsuario": {
															"type": "number",
															"format": "double"
														},
														"updatedAt": {
															"type": "string",
															"format": "date-time"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"perfilCompleto": {
															"type": "boolean"
														},
														"empresaId": {
															"type": "number",
															"format": "double"
														},
														"estado": {
															"type": "string"
														},
														"rolEquipo": {
															"$ref": "#/components/schemas/_36_Enums.rolEquipo"
														},
														"rol": {
															"type": "string"
														},
														"nit": {
															"type": "string"
														},
														"nombreEmpresa": {
															"type": "string"
														},
														"tipoUsuario": {
															"$ref": "#/components/schemas/_36_Enums.TipoUsuario"
														},
														"fotoPerfil": {
															"type": "string"
														},
														"direccion": {
															"type": "string"
														},
														"telefono": {
															"type": "string"
														},
														"nombreCompleto": {
															"type": "string"
														},
														"password": {
															"type": "string"
														},
														"correo": {
															"type": "string"
														},
														"username": {
															"type": "string"
														}
													},
													"required": [
														"deletedAt",
														"idUsuario",
														"updatedAt",
														"createdAt",
														"perfilCompleto",
														"empresaId",
														"estado",
														"rolEquipo",
														"rol",
														"nit",
														"nombreEmpresa",
														"tipoUsuario",
														"fotoPerfil",
														"direccion",
														"telefono",
														"nombreCompleto",
														"password",
														"correo",
														"username"
													],
													"type": "object"
												},
												"mensaje": {
													"type": "string"
												}
											},
											"required": [
												"eliminado",
												"mensaje"
											],
											"type": "object"
										},
										{
											"properties": {
												"eliminado": {},
												"mensaje": {}
											},
											"required": [
												"mensaje"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Equipo"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/equipo/todos/{empresaId}": {
			"delete": {
				"operationId": "EliminarTodoElEquipo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensaje": {
											"type": "string"
										}
									},
									"required": [
										"mensaje"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Equipo"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "empresaId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}